+++ lib/src/main/java/com/smileidentity/fragment/EnhancedSmartSelfieAuthenticationFragment.kt
@@ -7,11 +7,12 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.setFragmentResult
 import androidx.fragment.compose.content
 import com.smileidentity.SmileID
-import com.smileidentity.compose.SmartSelfieEnrollment
-import com.smileidentity.fragment.SmartSelfieEnrollmentFragment.Companion.KEY_REQUEST
-import com.smileidentity.fragment.SmartSelfieEnrollmentFragment.Companion.KEY_RESULT
-import com.smileidentity.fragment.SmartSelfieEnrollmentFragment.Companion.newInstance
-import com.smileidentity.fragment.SmartSelfieEnrollmentFragment.Companion.resultFromBundle
+import com.smileidentity.compose.SmartSelfieAuthentication
+import com.smileidentity.compose.SmartSelfieAuthenticationEnhanced
+import com.smileidentity.fragment.EnhancedSmartSelfieAuthenticationFragment.Companion.KEY_REQUEST
+import com.smileidentity.fragment.EnhancedSmartSelfieAuthenticationFragment.Companion.KEY_RESULT
+import com.smileidentity.fragment.EnhancedSmartSelfieAuthenticationFragment.Companion.newInstance
+import com.smileidentity.fragment.EnhancedSmartSelfieAuthenticationFragment.Companion.resultFromBundle
 import com.smileidentity.results.SmartSelfieResult
 import com.smileidentity.results.SmileIDResult
 import com.smileidentity.util.getParcelableCompat
@@ -21,11 +22,11 @@ import com.squareup.moshi.Types
 import kotlinx.collections.immutable.toImmutableMap
 
 /**
- * Perform a SmartSelfie™ Enrollment
+ * Perform a SmartSelfie™ Authentication
  *
  * [Docs](https://docs.usesmileid.com/products/for-individuals-kyc/biometric-authentication)
  *
- * A [Fragment] wrapper for the [SmartSelfieEnrollment] to be used if not using Jetpack
+ * A [Fragment] wrapper for the [SmartSelfieAuthentication] to be used if not using Jetpack
  * Compose. New instances *must* be created via [newInstance]. Results are communicated back to the
  * caller via [setFragmentResult]. Therefore, the caller must use
  * [androidx.fragment.app.FragmentManager.setFragmentResultListener] to listen for the result. If
@@ -38,13 +39,13 @@ import kotlinx.collections.immutable.toImmutableMap
  *
  * Usage example:
  * ```java
- * SmartSelfieEnrollmentFragment smartSelfieFragment = SmartSelfieEnrollmentFragment
+ * SmartSelfieAuthenticationFragment smartSelfieFragment = SmartSelfieAuthenticationFragment
  *  .newInstance();
  * getSupportFragmentManager().setFragmentResultListener(
- *   SmartSelfieEnrollmentFragment.KEY_REQUEST,
+ *   SmartSelfieAuthenticationFragment.KEY_REQUEST,
  *   this,
  *   (requestKey, result) -> {
- *     SmartSelfieResult smartSelfieResult = SmartSelfieEnrollmentFragment
+ *     SmartSelfieResult smartSelfieResult = SmartSelfieAuthenticationFragment
  *       .resultFromBundle(result);
  *     getSupportFragmentManager()
  *       .beginTransaction()
@@ -54,33 +55,34 @@ import kotlinx.collections.immutable.toImmutableMap
  *   );
  * ```
  */
-class SmartSelfieEnrollmentFragment : Fragment() {
+class EnhancedSmartSelfieAuthenticationFragment : Fragment() {
     companion object {
-        const val KEY_REQUEST = "SmartSelfieEnrollmentRequest"
+        const val KEY_REQUEST = "SmartSelfieAuthenticationRequest"
 
         /**
          * This is internal to prevent partners from accidentally using the wrong key. They only
          * need [KEY_REQUEST]. Partners should use [resultFromBundle] to extract the result.
          */
-        internal const val KEY_RESULT = "SmartSelfieEnrollmentResult"
+        internal const val KEY_RESULT = "SmartSelfieAuthenticationResult"
 
         /**
-         * Creates a new instance of [SmartSelfieEnrollmentFragment] which wraps the
-         * [SmileID.SmartSelfieEnrollment] Composable under the hood
+         * Creates a new instance of [EnhancedSmartSelfieAuthenticationFragment] which wraps the
+         * [SmileID.SmartSelfieAuthentication] Composable under the hood
          *
-         * @param userId The user ID to associate with the SmartSelfie™ Enrollment. Most often,
+         * @param userId The user ID to associate with the SmartSelfie™ Authentication. Most often,
          * this will correspond to a unique User ID within your own system. If not provided, a
          * random user ID will be generated.
-         * @param jobId The job ID to associate with the SmartSelfie™ Enrollment. Most often, this
+         * @param jobId The job ID to associate with the SmartSelfie™ Authentication. Most often, this
          * will correspond to a unique Job ID within your own system. If not provided, a random job ID
          * will be generated.
+         * @param allowNewEnroll Whether to allow new enrollments
          * @param allowAgentMode Whether to allow Agent Mode or not. If allowed, a switch will be
          * displayed allowing toggling between the back camera and front camera. If not allowed,
          * only the front camera will be used.
          * @param showAttribution Whether to show the Smile ID attribution or not.
-         * @param showInstructions Whether to deactivate capture screen's instructions for
+         * @param showInstructions Whether to show the instructions or not.
          * @param skipApiSubmission Whether to skip the API submission and return the result of capture only
-         * SmartSelfie.
+         * @param extraPartnerParams Custom values specific to partners
          */
         @JvmStatic
         @JvmOverloads
@@ -93,7 +95,7 @@ class SmartSelfieEnrollmentFragment : Fragment() {
             showInstructions: Boolean = true,
             skipApiSubmission: Boolean = false,
             extraPartnerParams: HashMap<String, String>? = null,
-        ) = SmartSelfieEnrollmentFragment().apply {
+        ) = EnhancedSmartSelfieAuthenticationFragment().apply {
             arguments = Bundle().apply {
                 this.userId = userId
                 this.jobId = jobId
@@ -116,11 +118,9 @@ class SmartSelfieEnrollmentFragment : Fragment() {
         savedInstanceState: Bundle?,
     ) = content {
         val args = requireArguments()
-        SmileID.SmartSelfieEnrollment(
+        SmileID.SmartSelfieAuthenticationEnhanced(
             userId = args.userId,
-            jobId = args.jobId,
             allowNewEnroll = args.allowNewEnroll,
-            allowAgentMode = args.allowAgentMode,
             showAttribution = args.showAttribution,
             showInstructions = args.showInstructions,
             skipApiSubmission = args.skipApiSubmission,
@@ -170,8 +170,11 @@ private var Bundle.skipApiSubmission: Boolean
     set(value) = putBoolean(KEY_SKIP_API_SUBMISSION, value)
 
 private const val KEY_EXTRA_PARTNER_PARAMS = "extraPartnerParams"
-private val type =
-    Types.newParameterizedType(Map::class.java, String::class.java, String::class.java)
+private val type = Types.newParameterizedType(
+    Map::class.java,
+    String::class.java,
+    String::class.java,
+)
 private val adapter = moshi.adapter<Map<String, String>>(type)
 private var Bundle.extraPartnerParams: Map<String, String>?
     get() = getString(KEY_EXTRA_PARTNER_PARAMS)?.let { adapter.fromJson(it) }