+++ lib/src/main/java/com/smileidentity/fragment/SmartSelfieAuthenticationFragment.kt
@@ -8,7 +8,6 @@ import androidx.fragment.app.setFragmentResult
 import androidx.fragment.compose.content
 import com.smileidentity.SmileID
 import com.smileidentity.compose.SmartSelfieAuthentication
-import com.smileidentity.compose.SmartSelfieAuthenticationEnhanced
 import com.smileidentity.fragment.SmartSelfieAuthenticationFragment.Companion.KEY_REQUEST
 import com.smileidentity.fragment.SmartSelfieAuthenticationFragment.Companion.KEY_RESULT
 import com.smileidentity.fragment.SmartSelfieAuthenticationFragment.Companion.newInstance
@@ -83,7 +82,6 @@ class SmartSelfieAuthenticationFragment : Fragment() {
          * only the front camera will be used.
          * @param showAttribution Whether to show the Smile ID attribution or not.
          * @param showInstructions Whether to show the instructions or not.
-         * @param useStrictMode Whether to use the enhanced version of the SmartSelfie™ Enrollment
          * @param skipApiSubmission Whether to skip the API submission and return the result of capture only
          * @param extraPartnerParams Custom values specific to partners
          */
@@ -96,7 +94,6 @@ class SmartSelfieAuthenticationFragment : Fragment() {
             allowAgentMode: Boolean = false,
             showAttribution: Boolean = true,
             showInstructions: Boolean = true,
-            useStrictMode: Boolean = false,
             skipApiSubmission: Boolean = false,
             extraPartnerParams: HashMap<String, String>? = null,
         ) = SmartSelfieAuthenticationFragment().apply {
@@ -107,7 +104,6 @@ class SmartSelfieAuthenticationFragment : Fragment() {
                 this.allowAgentMode = allowAgentMode
                 this.showAttribution = showAttribution
                 this.showInstructions = showInstructions
-                this.useStrictMode = useStrictMode
                 this.skipApiSubmission = skipApiSubmission
                 this.extraPartnerParams = extraPartnerParams
             }
@@ -123,34 +119,19 @@ class SmartSelfieAuthenticationFragment : Fragment() {
         savedInstanceState: Bundle?,
     ) = content {
         val args = requireArguments()
-        val useStrictMode = args.useStrictMode
-        if (useStrictMode) {
-            SmileID.SmartSelfieAuthenticationEnhanced(
-                userId = args.userId,
-                allowNewEnroll = args.allowNewEnroll,
-                showAttribution = args.showAttribution,
-                showInstructions = args.showInstructions,
-                skipApiSubmission = args.skipApiSubmission,
-                extraPartnerParams = (args.extraPartnerParams ?: mapOf()).toImmutableMap(),
-                onResult = {
-                    setFragmentResult(KEY_REQUEST, Bundle().apply { smileIdResult = it })
-                },
-            )
-        } else {
-            SmileID.SmartSelfieAuthentication(
-                userId = args.userId,
-                jobId = args.jobId,
-                allowNewEnroll = args.allowNewEnroll,
-                allowAgentMode = args.allowAgentMode,
-                showAttribution = args.showAttribution,
-                showInstructions = args.showInstructions,
-                skipApiSubmission = args.skipApiSubmission,
-                extraPartnerParams = (args.extraPartnerParams ?: mapOf()).toImmutableMap(),
-                onResult = {
-                    setFragmentResult(KEY_REQUEST, Bundle().apply { smileIdResult = it })
-                },
-            )
-        }
+        SmileID.SmartSelfieAuthentication(
+            userId = args.userId,
+            jobId = args.jobId,
+            allowNewEnroll = args.allowNewEnroll,
+            allowAgentMode = args.allowAgentMode,
+            showAttribution = args.showAttribution,
+            showInstructions = args.showInstructions,
+            skipApiSubmission = args.skipApiSubmission,
+            extraPartnerParams = (args.extraPartnerParams ?: mapOf()).toImmutableMap(),
+            onResult = {
+                setFragmentResult(KEY_REQUEST, Bundle().apply { smileIdResult = it })
+            },
+        )
     }
 }
 
@@ -185,11 +166,6 @@ private const val KEY_SHOW_INSTRUCTIONS = "showInstructions"
 private var Bundle.showInstructions: Boolean
     get() = getBoolean(KEY_SHOW_INSTRUCTIONS)
     set(value) = putBoolean(KEY_SHOW_INSTRUCTIONS, value)
-
-private const val KEY_USE_STRICT_MODE = "useStrictMode"
-private var Bundle.useStrictMode: Boolean
-    get() = getBoolean(KEY_USE_STRICT_MODE)
-    set(value) = putBoolean(KEY_USE_STRICT_MODE, value)
 
 private const val KEY_SKIP_API_SUBMISSION = "skipApiSubmission"
 private var Bundle.skipApiSubmission: Boolean